added the treeview
Treeview added.

The number besides the directory, is the number of .gpg files in it.

You can probably combine the ListDirectory() and FillDirectoryTree() but
i thought you'd want to do that yourself

I added an event = toolStriptextSearch_Enter
When you click on the search bar, it changes the datatable to the one,
with all the entries in it (variable dt, whereas the treeview uses
treeDt)

And when you double click a node on the tree, it changes back